+++ releaf-core/lib/generators/dummy/templates/controllers/admin/chapters_controller.rb
@@ -1,3 +1,3 @@
-class Admin::ChaptersController < Releaf::BaseController
+class Admin::ChaptersController < Releaf::ActionController
 end