+++ releaf-permissions/app/controllers/releaf/permissions/users_controller.rb
@@ -1,19 +1,16 @@
-module Releaf::Permissions
-  class UsersController < Releaf::BaseController
+class Releaf::Permissions::UsersController < Releaf::ActionController
+  def self.resource_class
+    Releaf::Permissions::User
+  end
 
-    def self.resource_class
-      Releaf::Permissions::User
-    end
+  protected
 
-    protected
+  def prepare_new
+    super
+    @resource.role = Releaf::Permissions::Role.first
+  end
 
-    def prepare_new
-      super
-      @resource.role = Releaf::Permissions::Role.first
-    end
-
-    def permitted_params
-      %w[name surname role_id email password password_confirmation locale]
-    end
+  def permitted_params
+    %w[name surname role_id email password password_confirmation locale]
   end
 end